Updated with possibility to have properties and values

+++ librave/toolbox/rave_value.c
@@ -278,7 +278,7 @@ int RaveValue_getDouble(RaveValue_t* self, double* value)
   return result;
 }
 
-long RaveValue_toDouble(RaveValue_t* self)
+double RaveValue_toDouble(RaveValue_t* self)
 {
   RAVE_ASSERT((self != NULL), "attr == NULL");
   return self->doubleValue;
